setup/requirements.txt
boto3==1.33.1
botocore==1.33.1
cffi==1.16.0
cryptography==41.0.7
DateTime==5.3
jmespath==1.0.1
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.2
pandas==2.1.3
protobuf==4.21.12
psycopg2==2.9.9
pyaes==1.6.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.8.0
schedule==1.2.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
zope.interface==6.1
